+++ lightly_studio/src/lightly_studio/examples/example_operators.py
@@ -105,7 +105,7 @@ for i in range(20):
 dataset_path = env.path("EXAMPLES_DATASET_PATH", "/path/to/your/dataset")
 
 # Create a DatasetLoader from a path
-dataset = ls.Dataset.create()
+dataset = ls.ImageDataset.create()
 dataset.add_images_from_path(path=dataset_path)
 
 ls.start_gui()