+++ lightly_studio/src/lightly_studio/examples/example_yolo.py
@@ -17,6 +17,6 @@ dataset_path = env.path("EXAMPLES_YOLO_YAML_PATH", "/path/to/your/dataset/data.y
 input_split = env.str("EXAMPLES_YOLO_SPLIT", "train")
 
 # Create a DatasetLoader from a path
-dataset = ls.Dataset.create()
+dataset = ls.ImageDataset.load()
 dataset.add_samples_from_yolo(data_yaml=dataset_path, input_split=input_split)
 ls.start_gui()