+++ lightly_studio/tests/core/dataset_query/test_tags_expression.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from sqlmodel import Session, select
 
-from lightly_studio.core.dataset_query.sample_field import SampleField
+from lightly_studio.core.dataset_query.image_sample_field import ImageSampleField
 from lightly_studio.models.image import ImageTable
 from lightly_studio.resolvers import tag_resolver
 from tests.helpers_resolvers import create_collection, create_image, create_tag
@@ -10,13 +10,13 @@ from tests.helpers_resolvers import create_collection, create_image, create_tag
 
 class TestTagsContainsExpression:
     def test_apply(self) -> None:
-        expr = SampleField.tags.contains("car")
+        expr = ImageSampleField.tags.contains("car")
         assert expr.tag_name == "car"
 
     def test_apply__sql(self) -> None:
         """Test that TagsContainsExpression correctly modifies the SQL query."""
         query = select(ImageTable)
-        query = query.where(SampleField.tags.contains("car").get())
+        query = query.where(ImageSampleField.tags.contains("car").get())
         sql = str(query.compile(compile_kwargs={"literal_binds": True}))
 
         # The current approach makes a subquery for the tags relationship.
@@ -32,8 +32,8 @@ class TestTagsContainsExpression:
         tag_resolver.add_tag_to_sample(session=test_db, tag_id=tag.tag_id, sample=image.sample)
 
         query = select(ImageTable)
-        query = query.where(SampleField.tags.contains("vehicle").get())
-        query = query.where(SampleField.tags.contains("car").get())
+        query = query.where(ImageSampleField.tags.contains("vehicle").get())
+        query = query.where(ImageSampleField.tags.contains("car").get())
 
         # The sample has only one out of the two tags, no results are expected
         results = test_db.exec(query).all()